+++ lib/karafka/errors.rb
@@ -31,16 +31,6 @@ module Karafka
     # (and we couldn't find) any appropriate responder for a given controller
     class ResponderMissing < BaseError; end
 
-    # Raised when we want to use #respond_to in responders with a topic that we didn't register
-    class UnregisteredTopic < BaseError; end
-
-    # Raised when we send more than one message to a single topic but we didn't allow that when
-    # we were registering topic in a responder
-    class TopicMultipleUsage < BaseError; end
-
-    # Raised when we didn't use a topic that was defined as non-optional (required)
-    class UnusedResponderRequiredTopic < BaseError; end
-
     # Raised when configuration doesn't match with validation schema
     class InvalidConfiguration < BaseError; end
   end